requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2
psycopg[binary]
pgvector
redis
pypdf
google-generativeai
langgraph
python-dotenv
python-multipart
google-genai
